+++ spiel/widgets/widget.py
@@ -5,6 +5,8 @@ from typing import TYPE_CHECKING
 from textual.reactive import watch
 from textual.widget import Widget
 
+from spiel.slide import Slide
+
 if TYPE_CHECKING:
     from spiel.app import SpielApp
 
@@ -19,3 +21,7 @@ class SpielWidget(Widget):
 
     def r(self, _: object) -> None:
         self.refresh()
+
+    @property
+    def current_slide(self) -> Slide:
+        return self.app.deck[self.app.current_slide_idx]