+++ docs/changelog.md
@@ -4,6 +4,14 @@
 
 ## `0.4.5` | *Unreleased*
 
+### Fixed
+
+- [#202](https://github.com/JoshKarpel/spiel/pull/202) Returning un-renderable content from a slide content function now displays an error instead of crashing Spiel.
+
+### Changed
+
+- [#203](https://github.com/JoshKarpel/spiel/pull/203) The `Image` example in the demo deck is now centered inside its `Panel`.
+
 ## `0.4.4` | 2023-01-13
 
 ### Added